part 3 stuff
all of the mat files are model fits with the Y_fit, x, etc variables

the "x(5) constriant" one has it so that the breakthrough infection rate has to be less than or equal to the susceptible infection rate. Not sure whether or not that is a valid assumtion to make -- i.e. do the parameters in the matrix really line up to true infection rates, or are we interpriting them incorrectly?